Fix/tests (#50)

+++ internal/application/service/pet_service.go
@@ -39,14 +39,6 @@ func (ps *PetService) AddPet(info pet.AddPetRequest) error {
 	}
 
 	info.IsAdult = isAdult
-	var profileKey *string
-	if info.ProfilePic != nil {
-		profileKeyValue := ps.getStorageKey(info.Name, info.UserID)
-		profileKey = &profileKeyValue
-		if err := ps.storage.UploadFile(enums.PetProfilePic, *profileKey, info.ProfilePic); err != nil {
-			return err
-		}
-	}
 
 	_, err = ps.findPet(info.Name, info.UserID)
 	if err == nil {
@@ -55,6 +47,16 @@ func (ps *PetService) AddPet(info pet.AddPetRequest) error {
 		return &ce
 	} else if _, ok := err.(*exceptions.NotFoundError); !ok {
 		return err
+	}
+
+	var profileKey *string
+	if info.ProfilePic != nil {
+		log.Println()
+		profileKeyValue := ps.getStorageKey(info.Name, info.UserID)
+		profileKey = &profileKeyValue
+		if err := ps.storage.UploadFile(enums.PetProfilePic, *profileKey, info.ProfilePic); err != nil {
+			return err
+		}
 	}
 
 	pet := &entities.Pet{
@@ -82,6 +84,10 @@ func (ps *PetService) UpdatePet(info pet.UpdatePetRequest) error {
 	foundPet, err := ps.FindPetByID(info.ID)
 	if err != nil {
 		return err
+	}
+
+	if foundPet.UserID != info.UserID {
+		return exceptions.NewAccessDeniedError("can't update other's pets")
 	}
 
 	err = ps.validateSpecies(info.Species, info.Kind)
@@ -136,6 +142,9 @@ func (ps *PetService) RemovePet(info pet.RemovePetRequest) error {
 	foundPet, err := ps.FindPetByID(info.ID)
 	if err != nil {
 		return err
+	}
+	if foundPet.UserID != info.UserID {
+		return exceptions.NewAccessDeniedError("can't delete other's pets!")
 	}
 	profileKeyValue := ps.getStorageKey(foundPet.Name, foundPet.UserID)
 	if err = ps.storage.DeleteObject(enums.PetProfilePic, profileKeyValue); err != nil {
@@ -294,12 +303,12 @@ func (ps *PetService) validateSpecies(species enums.Species, kind enums.PetKind)
 
 func (ps *PetService) GetAllPetKinds() []pet.PetKindResponse {
 	kinds := enums.GetAllPetKinds()
-	res := make([]pet.PetKindResponse, 0)
-	for _, kind := range kinds {
-		res = append(res, pet.PetKindResponse{
+	res := make([]pet.PetKindResponse, len(kinds))
+	for i, kind := range kinds {
+		res[i] = pet.PetKindResponse{
 			Num:  kind,
 			Name: kind.String(),
-		})
+		}
 	}
 	return res
 }
@@ -307,25 +316,25 @@ func (ps *PetService) GetAllPetKinds() []pet.PetKindResponse {
 func (ps *PetService) GetPetKindSpecies(info pet.GetPetKindSpecies) []pet.PetSpeciesResponse {
 	species := enums.GetSpeciesByKind(info.Num)
 	species = append(species, enums.Other)
-	res := make([]pet.PetSpeciesResponse, 0)
-	for _, s := range species {
-		res = append(res, pet.PetSpeciesResponse{
+	res := make([]pet.PetSpeciesResponse, len(species))
+	for i, s := range species {
+		res[i] = pet.PetSpeciesResponse{
 			Num:  s,
 			Name: s.String(),
-		})
+		}
 	}
 	return res
 }
 
 func (ps *PetService) GetPetsBasicDataResponse(pets []entities.Pet) ([]pet.PetBasicDataResponse, error) {
-	r := make([]pet.PetBasicDataResponse, 0)
+	r := make([]pet.PetBasicDataResponse, len(pets))
 
-	for _, petEntity := range pets {
+	for i, petEntity := range pets {
 		res, err := ps.getPetBasicDataResponse(&petEntity)
 		if err != nil {
 			return nil, err
 		}
-		r = append(r, *res)
+		r[i] = *res
 	}
 
 	return r, nil
@@ -353,9 +362,9 @@ func (ps *PetService) GetPetsInUser(userPets []entities.Pet, petIDs []uint) ([]e
 }
 
 func (ps *PetService) GetPetNames(pets []entities.Pet) []string {
-	names := make([]string, 0)
-	for _, pet := range pets {
-		names = append(names, pet.Name)
+	names := make([]string, len(pets))
+	for i, pet := range pets {
+		names[i] = pet.Name
 	}
 
 	return names
